update: comments, creator

+++ app/Http/Controllers/Api/Order/OrderGetController.php
@@ -25,7 +25,7 @@ class OrderGetController extends Controller
     public function getOrder($orderId)
     {
         try {
-            $order = Order::with('order_items', 'contributions')->findOrFail($orderId);
+            $order = Order::with('order_items', 'contributions')->whereHas('order_items')->findOrFail($orderId);
 
             $this->authorize('getOrder', $order);
 
@@ -35,7 +35,7 @@ class OrderGetController extends Controller
             $carbonDate = Carbon::parse($order->updated_at);
             $order->time_ago = $carbonDate->diffForHumans();
 
-            foreach($order->order_items as $item){
+            foreach ($order->order_items as $item) {
                 $item->slug_name =  Str::slug($item->product()->select('title')->first()->title);
                 $item->temporal_price = $item->product()->select('current_price')->first()->current_price;
             }
@@ -44,7 +44,6 @@ class OrderGetController extends Controller
                 'status' => 'success',
                 'data' => $order
             ], 200);
-
         } catch (AuthorizationException $e) {
             return response()->json([
                 'status' => 'error',
@@ -80,7 +79,7 @@ class OrderGetController extends Controller
                     $carbonDate = Carbon::parse($order->updated_at);
                     $order->time_ago = $carbonDate->diffForHumans();
                 }
-                
+
                 return response()->json([
                     'status' => 'success',
                     'current_page' => $orders->currentPage(),
@@ -185,17 +184,27 @@ class OrderGetController extends Controller
     }
 
 
-    public function delete(Order $order)
+    public function delete($id)
     {
         try {
+            $order = Order::withTrashed()->find($id);
+
+            if (!$order || $order->trashed()) {
+                return response()->json([
+                    'status' => 'success',
+                    'message' => 'La commande a déjà été supprimée.'
+                ], 200);
+            }
+
             $this->authorize('delete', $order);
-            
+
             $order->delete();
 
             return response()->json([
                 'status' => 'success',
                 'message' => 'Suppression effectuée avec succès.'
             ], 200);
+            
         } catch (AuthorizationException $e) {
             return response()->json([
                 'status' => 'error',
